Add other preview for discord

+++ index.js
@@ -2,45 +2,54 @@ const http = require('http');
 const express = require('express');
 const Jimp = require('jimp');
 const fs = require('fs');
-const CronJob = require('cron').CronJob;
 
 const app = express();
 app.set('trust proxy', 1)
 
 app.get('/mario:ext', (req, res) => {
-    // get ip
-    let ip = req.headers['x-forwarded-for'] || req.socket.remoteAddress;
-    ip = ip.split(',')[0];
 
-    if (ip == "::1") // localhost
-        ip = "192.0.2.0";
+    if (req.headers['user-agent'] === "Mozilla/5.0 (Macintosh; Intel Mac OS X 11.6; rv:92.0) Gecko/20100101 Firefox/92.0") {
+        res.sendFile(`${__dirname}/normal.png`);
+    } else {
 
-    // create image
-    let path = `${__dirname}/tmp/${ip}.jpg`;
-    fs.access(path, fs.F_OK, (err) => {
-        if (err) { // file does not exist
-            createImg(ip, (err, imgpath) => {
-                if (err) throw err;
-                res.sendFile(path);
-            });
-            return
-        }
+        // get ip
+        console.log(req.headers)
+        let ip = req.headers['x-forwarded-for'] || req.socket.remoteAddress;
+        ip = ip.split(',')[0];
 
-        res.sendFile(path); // file already exists
-    });
+        if (ip === "::1") // localhost
+            ip = "192.0.2.0";
 
-    // delete the file after 10 minutes
-    setTimeout(() => {
+        // clean file path
+        let cleanIp = ip.replace(/:/g, '_');
+
+        // create image
+        let path = `${__dirname}/tmp/${cleanIp}.jpg`;
         fs.access(path, fs.F_OK, (err) => {
-            if (err) return;
-            fs.unlink(path, (err) => {
-                if (err) throw err;
-            });
+            if (err) { // file does not exist
+                createImg(ip, path, (err, path) => {
+                    if (err) throw err;
+                    res.sendFile(path);
+                });
+                return
+            }
+
+            res.sendFile(path); // file already exists
         });
-    }, 1000 * 60 * 10);
+
+        // delete the file after 10 minutes
+        setTimeout(() => {
+            fs.access(path, fs.F_OK, (err) => {
+                if (err) return;
+                fs.unlink(path, (err) => {
+                    if (err) throw err;
+                });
+            });
+        }, 1000 * 60 * 10);
+    }
 });
 
-const port = 3000;
+const port = 3001;
 const server = http.createServer(app);
 
 server.listen(port, () => {
@@ -50,11 +59,11 @@ server.listen(port, () => {
 /**
  * Creates a mario image and returns it
  * @param {string} ip The IP to add to the image
+ * @param {string} path The path to save the image to
  * @param {callback} callback The callback to call when done
  */
-function createImg(ip, callback) {
+function createImg(ip, path, callback) {
     let img;
-    let path = `tmp/${ip}.jpg`;
 
     if (!fs.existsSync("tmp")) {
         fs.mkdirSync("tmp");
